pyproject.toml
[project]
name = "ai-optimizers-benchmark"
version = "1.0.0"
description = "Publication-quality benchmarking suite comparing Adam, AdaMuon, and K-FAC optimizers"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "nvidia-ml-py>=12.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
